[MOV] html_editor,web_editor: move tools.py to html_editor
