+++ addons/hr/models/hr_job.py
@@ -2,7 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import api, fields, models, _
-from odoo.addons.web_editor.tools import handle_history_divergence
+from odoo.addons.html_editor.tools import handle_history_divergence
 
 
 class HrJob(models.Model):
