+++ addons/web_editor/tests/__init__.py
@@ -3,4 +3,3 @@
 
 from . import test_controller
 from . import test_odoo_editor
-from . import test_tools
