+++ addons/website_sale/controllers/main.py
@@ -19,7 +19,7 @@ from odoo.tools.translate import LazyTranslate, _
 
 from odoo.addons.payment.controllers import portal as payment_portal
 from odoo.addons.sale.controllers import portal as sale_portal
-from odoo.addons.web_editor.tools import get_video_thumbnail
+from odoo.addons.html_editor.tools import get_video_thumbnail
 from odoo.addons.website.controllers.main import QueryURL
 from odoo.addons.website.models.ir_http import sitemap_qs2dom
 from odoo.addons.website_sale.const import SHOP_PATH
