+++ addons/website_sale/models/product_image.py
@@ -6,7 +6,7 @@ from odoo import _, api, fields, models
 from odoo.exceptions import ValidationError
 from odoo.tools.image import is_image_size_above
 
-from odoo.addons.web_editor.tools import get_video_embed_code, get_video_thumbnail
+from odoo.addons.html_editor.tools import get_video_embed_code, get_video_thumbnail
 
 
 class ProductImage(models.Model):
